% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_corpus_moviereviews}
\alias{data_corpus_moviereviews}
\title{Movie reviews from Pang and Lee (2004)}
\source{
\url{https://www.cs.cornell.edu/people/pabo/movie-review-data/}
}
\description{
A corpus object containing 2,000 movie reviews.
}
\references{
Pang, B., Lee, L.  (2004)
"\href{https://www.cs.cornell.edu/home/llee/papers/cutsent.pdf}{A Sentimental Education: Sentiment Analysis Using Subjectivity Summarization Based on Minimum Cuts.}", Proceedings of the ACL.
}
