% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divergence}
\alias{divergence}
\title{Optimize the number of topics}
\usage{
divergence(x)
}
\arguments{
\item{x}{a LDA model fitted by \code{\link[=textmodel_seededlda]{textmodel_seededlda()}} or \code{\link[=textmodel_lda]{textmodel_lda()}}}
}
\description{
\code{divergence()} computes the divergence of topics to help users to find the
optimal number of topics for LDA. Divergence is the average Kullback–Leibler
distance between all the pairs of topic vectors in \code{x$phi}. The divergence
score maximizes when the chosen number of topic \code{k} is optimal (Deveaud et
al., 2014).
}
\references{
Deveaud, Romain et al. (2014). "Accurate and Effective Latent
Concept Modeling for Ad Hoc Information Retrieval".
doi:10.3166/DN.17.1.61-84. \emph{Document Numérique}.
}
