% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seededlda.R
\name{print.textmodel_lda}
\alias{print.textmodel_lda}
\title{Print method for a LDA model}
\usage{
\method{print}{textmodel_lda}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed.}

\item{...}{unused.}
}
\description{
Print method for a LDA model
}
\keyword{internal}
