% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sizes}
\alias{sizes}
\title{[experimental] Compute the sizes of topics}
\usage{
sizes(x)
}
\arguments{
\item{x}{a LDA model fitted by \code{\link[=textmodel_seededlda]{textmodel_seededlda()}} or \code{\link[=textmodel_lda]{textmodel_lda()}}}
}
\description{
Compute the sizes of topics as the proportions of topic words in the corpus.
}
