% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seededlda.R
\name{textmodel_seededlda}
\alias{textmodel_seededlda}
\title{Fit seeded LDA with dictionary}
\usage{
textmodel_seededlda(x, dictionary, weight = 500, residual = FALSE, ...)
}
\arguments{
\item{x}{\code{dfm}}

\item{dictionary}{a dictionary}

\item{weight}{pseudo-count given to words in dictionary to adjust prior
probability of topics over words}

\item{residual}{if \code{TRUE} a residual topic will be added to user-defined
topics for documents that do not fall into any of them}

\item{...}{additional arguments to specify how to interpret the dictionary}
}
\description{
Fit seeded LDA with dictionary
}
