% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{textmodel_sequentiallda}
\alias{textmodel_sequentiallda}
\title{Unsupervised Sequential Latent Dirichlet allocation}
\usage{
textmodel_sequentiallda(
  x,
  k = 10,
  max_iter = 2000,
  alpha = 0.5,
  beta = 0.1,
  model = NULL,
  verbose = quanteda_options("verbose")
)
}
\description{
\code{textmodel_sequentiallda()} is a shortcut function for
Sequential LDA, equivalent to \code{textmodel_lda(gamma = 0.5)}.
}
