% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seededlda.R
\name{tfm}
\alias{tfm}
\title{Internal function to construct topic-feature matrix}
\usage{
tfm(x, dictionary, levels = 1:5, valuetype = c("glob", "regex",
  "fixed"), case_insensitive = TRUE, weight = 500,
  scheme = c("absolute", "relative"), residual = TRUE)
}
\description{
Internal function to construct topic-feature matrix
}
