% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.dictionary.textmodel_lda}
\alias{as.dictionary.textmodel_lda}
\title{Create a dictionary from topic terms}
\usage{
\method{as.dictionary}{textmodel_lda}(x, separator = NULL, n = 10, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link[=textmodel_lda]{textmodel_lda()}}.}

\item{separator}{the character in between multi-word dictionary values.}

\item{n}{the number of terms in the dictionary for each topic.}

\item{...}{not used.}
}
\value{
Returns a \link[quanteda:dictionary]{quanteda::dictionary} object.
}
\description{
\code{as.dictionary()} creates a \link[quanteda:dictionary]{quanteda::dictionary} object from top topic terms.
}
