% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perplexity}
\alias{perplexity}
\title{Optimize the hyper-parameters}
\usage{
perplexity(x, newdata = NULL, max_iter = 100)
}
\arguments{
\item{x}{the dfm on which the model will be fit.}

\item{newdata}{the dfm for which the perplexity score will be computed. If \code{NULL}, \code{x$data} will be used.}

\item{max_iter}{the maximum number of iteration in Gibbs sampling.}
}
\description{
\code{perplexity()} computes the perplexity score of a fitted LDA model to optimize hyper-parameters.
}
