% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seededlda.R
\name{terms}
\alias{terms}
\title{Extract most likely terms}
\usage{
terms(x, k = 10)
}
\arguments{
\item{x}{a fitted seeded LDA model}

\item{k}{number of terms to be extracted}
}
\description{
Extract most likely terms
}
