% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R, R/seededlda.R
\name{textmodel_lda}
\alias{textmodel_lda}
\alias{textmodel_seededlda}
\title{Semisupervised Latent Dirichlet allocation}
\usage{
textmodel_lda(
  x,
  k = 10,
  max_iter = 2000,
  alpha = NULL,
  beta = NULL,
  verbose = quanteda_options("verbose")
)

textmodel_seededlda(
  x,
  dictionary,
  valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE,
  residual = FALSE,
  weight = 0.01,
  max_iter = 2000,
  alpha = NULL,
  beta = NULL,
  verbose = quanteda_options("verbose")
)
}
\arguments{
\item{x}{the dfm on which the model will be fit}

\item{k}{the number of topics}

\item{max_iter}{the maximum number of iteration in Gibbs sampling.}

\item{alpha}{the hyper parameter for topic-document distribution}

\item{beta}{the hyper parameter for topic-word distribution}

\item{verbose}{logical; if `TRUE` print diagnostic information during
fitting.}

\item{dictionary}{a [dictionary][quanteda::dictionary] with seed words as
examples of topics.}

\item{valuetype}{see [valuetype][quanteda::valuetype]}

\item{case_insensitive}{see [valuetype][quanteda::valuetype]}

\item{residual}{if \code{TRUE} a residual topic (or "garbage topic") will be
added to user-defined topics.}

\item{weight}{pseudo count given to seed words as a proportion of total
number of words in `x`.}
}
\description{
`textmodel_seededlda()` implements semisupervised Latent Dirichlet allocation (seeded-LDA).
The estimator's code adopted from the GibbsLDA++ library (Xuan-Hieu Phan,
2007). `textmodel_seededlda()` allows identification of pre-defined topics by
semisupervised learning with a seed word dictionary.
}
\references{
Lu, Bin et al. (2011).
  [Multi-aspect Sentiment Analysis with Topic Models](https://dl.acm.org/doi/10.5555/2117693.2119585).
  *Proceedings of the 2011 IEEE 11th International Conference on Data Mining Workshops*.

  Watanabe, Kohei & Zhou, Yuan (2020).
  [Theory-Driven Analysis of Large Corpora: Semisupervised Topic Classification of the UN Speeches](https://doi.org/10.1177/0894439320907027).
  *Social Science Computer Review*.
}
\seealso{
[topicmodels][topicmodels::LDA]
}
\keyword{experimental}
\keyword{textmodel}
